Allow character device files to be used as log file
By allowing to set LOG_FILE to a character device file (such as /dev/stdout or /dev/stderr), we provide users more control over logging.
This is especially useful when running Lemur is a containerized where it is generally regarded as a best practice to send all logs to stdout/stderr.