Add scaffolding for Azure source plugin (#24)
The Azure plugin code organization assumes that only a destination plugin implementation will exist. This is a somewhat non-ideal assumption because an Azure source plugin will eventually be required for discovering endpoints and certificates within a given tenant. While a future source plugin for Azure is a nice to have capability for observability reasons and not required for automating certificate rotations, it seems to make sense that the code be re-organized to cleanly accommodate both a source and destination plugin implementation before proceeding further with any migrations.

Relates to [EDGE-1725](https://datadoghq.atlassian.net/browse/EDGE-1725).